+++ deepparse/network/seq2seq.py
@@ -1,8 +1,5 @@
 # pylint: disable=too-many-arguments
 
-# Pylint raise error for torch.tensor, torch.zeros, ... as a no-member event
-# if not the case.
-# pylint: disable=no-member
 import os
 import random
 import warnings