+++ deepparse/tools.py
@@ -82,18 +82,10 @@ def latest_version(model: str, cache_path: str, verbose: bool) -> bool:
     return is_latest_version
 
 
-def download_from_url(file_name: str, saving_dir: str, file_extension: str) -> None:  # pragma: no cover
-    warnings.warn(
-        "download_from_url is deprecated; use download_from_public_repository to download files from "
-        "our public repository. The function will be removed in the next major release."
-    )
-    download_from_public_repository(file_name=file_name, saving_dir=saving_dir, file_extension=file_extension)
-
-
 def download_from_public_repository(file_name: str, saving_dir: str, file_extension: str) -> None:
     """
     Simple function to download the content of a file from Deepparse public repository.
-    The repository URL string is  ̀`'https://graal.ift.ulaval.ca/public/deepparse/{}.{}'``
+    The repository URL string is  `'https://graal.ift.ulaval.ca/public/deepparse/{}.{}'``
     where the first bracket is the file name and the second is the file extension.
     """
     url = BASE_URL.format(file_name, file_extension)