+++ examples/retrain_with_new_prediction_tags.py
@@ -6,7 +6,7 @@ from deepparse import download_from_public_repository
 from deepparse.dataset_container import PickleDatasetContainer
 from deepparse.parser import AddressParser
 
-# First, let's download the train and test data with "new tags" from the public repository.
+# First, let's download the train and test data with the new tags, "new tags", from the public repository.
 saving_dir = "./data"
 file_extension = "p"
 training_dataset_name = "sample_incomplete_data_new_prediction_tags"
@@ -18,7 +18,7 @@ download_from_public_repository(test_dataset_name, saving_dir, file_extension=fi
 training_container = PickleDatasetContainer(os.path.join(saving_dir, training_dataset_name + "." + file_extension))
 test_container = PickleDatasetContainer(os.path.join(saving_dir, test_dataset_name + "." + file_extension))
 
-# We will retrain the fasttext version of our pretrained model.
+# We will retrain the FastText version of our pretrained model.
 model = "fasttext"
 address_parser = AddressParser(model_type=model, device=0)
 
@@ -35,7 +35,7 @@ logging_path = "./checkpoints"
 
 address_parser.retrain(
     training_container,
-    0.8,
+    train_ratio=0.8,
     epochs=5,
     batch_size=8,
     num_workers=2,
@@ -51,7 +51,7 @@ address_parser.test(test_container, batch_size=256)
 # That way, training will be faster.
 address_parser.retrain(
     training_container,
-    0.8,
+    train_ratio=0.8,
     epochs=5,
     batch_size=8,
     num_workers=2,