+++ tests/network/test_fasttext_seq2seq_model_cpu.py
@@ -4,10 +4,6 @@
 # Bug with PyTorch source code makes torch.tensor as not callable for pylint.
 # pylint: disable=not-callable
 
-# Pylint raise error for torch.tensor, torch.zeros, ... as a no-member event
-# if not the case.
-# pylint: disable=no-member
-
 # Pylint raise error for the call method mocking
 # pylint: disable=unnecessary-dunder-call