+++ tests/parser/integration/test_integration_address_parser_test_cpu.py
@@ -15,19 +15,18 @@ from tests.parser.integration.base_retrain import AddressParserRetrainTestCase
     not os.path.exists(os.path.join(os.path.expanduser("~"), ".cache", "deepparse", "cc.fr.300.bin")),
     "download of model too long for test in runner",
 )
-class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
+class AddressParserIntegrationTestAPICPUTest(AddressParserRetrainTestCase):
     def test_givenAFasttextAddressParser_whenTestWithNumWorkerAt0_thenTestOccur(self):
         address_parser = AddressParser(
             model_type=self.a_fasttext_model_type,
             device=self.a_cpu_device,
             verbose=self.verbose,
         )
-        self.training(address_parser, self.training_container, self.a_zero_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
-            num_workers=self.a_number_of_workers,
+            num_workers=self.a_zero_number_of_workers,
         )
 
         self.assertIsNotNone(performance_after_test)
@@ -40,7 +39,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_cpu_device,
             verbose=self.verbose,
         )
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,
@@ -59,8 +57,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             verbose=self.verbose,
         )
 
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
-
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
@@ -75,8 +71,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_cpu_device,
             verbose=self.verbose,
         )
-
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,
@@ -94,8 +88,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_cpu_device,
             verbose=self.verbose,
         )
-
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         callback_mock = MagicMock()
         performance_after_test = address_parser.test(
@@ -128,8 +120,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             verbose=self.verbose,
         )
 
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
-
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
@@ -145,8 +135,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             verbose=self.verbose,
         )
 
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
-
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
@@ -161,12 +149,11 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_cpu_device,
             verbose=self.verbose,
         )
-        self.training(address_parser, self.training_container, self.a_zero_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
-            num_workers=self.a_number_of_workers,
+            num_workers=self.a_zero_number_of_workers,
         )
 
         self.assertIsNotNone(performance_after_test)
@@ -179,7 +166,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_cpu_device,
             verbose=self.verbose,
         )
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,
@@ -198,8 +184,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             verbose=self.verbose,
         )
 
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
-
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
@@ -214,8 +198,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_cpu_device,
             verbose=self.verbose,
         )
-
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,
@@ -233,8 +215,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_cpu_device,
             verbose=self.verbose,
         )
-
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         callback_mock = MagicMock()
         performance_after_test = address_parser.test(
@@ -267,8 +247,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             verbose=self.verbose,
         )
 
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
-
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
@@ -283,8 +261,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_cpu_device,
             verbose=self.verbose,
         )
-
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,