+++ tests/parser/integration/test_integration_address_parser_test_gpu.py
@@ -12,19 +12,18 @@ from tests.parser.integration.base_retrain import AddressParserRetrainTestCase
 
 
 @skipIf(not torch.cuda.is_available(), "no gpu available")
-class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
+class AddressParserIntegrationTestAPIGPUTest(AddressParserRetrainTestCase):
     def test_givenAFasttextAddressParser_whenTestWithNumWorkerAt0_thenTestOccur(self):
         address_parser = AddressParser(
             model_type=self.a_fasttext_model_type,
             device=self.a_torch_device,
             verbose=self.verbose,
         )
-        self.training(address_parser, self.training_container, self.a_zero_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
-            num_workers=self.a_number_of_workers,
+            num_workers=self.a_zero_number_of_workers,
         )
 
         self.assertIsNotNone(performance_after_test)
@@ -37,7 +36,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_torch_device,
             verbose=self.verbose,
         )
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,
@@ -56,8 +54,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             verbose=self.verbose,
         )
 
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
-
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
@@ -72,8 +68,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_torch_device,
             verbose=self.verbose,
         )
-
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,
@@ -91,8 +85,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_torch_device,
             verbose=self.verbose,
         )
-
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         callback_mock = MagicMock()
         performance_after_test = address_parser.test(
@@ -125,8 +117,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             verbose=self.verbose,
         )
 
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
-
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
@@ -142,8 +132,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             verbose=self.verbose,
         )
 
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
-
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
@@ -158,12 +146,11 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_torch_device,
             verbose=self.verbose,
         )
-        self.training(address_parser, self.training_container, self.a_zero_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
-            num_workers=self.a_number_of_workers,
+            num_workers=self.a_zero_number_of_workers,
         )
 
         self.assertIsNotNone(performance_after_test)
@@ -176,8 +163,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_torch_device,
             verbose=self.verbose,
         )
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
-
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
@@ -195,8 +180,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             verbose=self.verbose,
         )
 
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
-
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
@@ -211,8 +194,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_torch_device,
             verbose=self.verbose,
         )
-
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,
@@ -230,8 +211,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_torch_device,
             verbose=self.verbose,
         )
-
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         callback_mock = MagicMock()
         performance_after_test = address_parser.test(
@@ -264,8 +243,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             verbose=self.verbose,
         )
 
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
-
         performance_after_test = address_parser.test(
             self.test_container,
             batch_size=self.a_batch_size,
@@ -280,8 +257,6 @@ class AddressParserIntegrationTestAPITest(AddressParserRetrainTestCase):
             device=self.a_torch_device,
             verbose=self.verbose,
         )
-
-        self.training(address_parser, self.training_container, self.a_number_of_workers)
 
         performance_after_test = address_parser.test(
             self.test_container,